Add serde support
This will help solve #179. The main challenge is to serialize the start time reference
of the rate limiter state.

I've gone for the naive approach of approximating the start time using the system time,
this is not very precise and can be affected by:
- Clock drift
- Timezone changes
- System time modifications
- System suspension under some operating systems.

The use-case is to be able to reload the rate limiter state from disk, if the clock is moved then
we saturate to 0, so at worst the rate limiter will take longer to regain tokens.